third_party/WebKit/LayoutTests/fast/dom/custom/registration-context-delete-during-register-base-constructor-retrieval.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script src="testutils.js"></script>
<body>
<script>
if (fork()) {
    // The controlling parent frame
    description('Tests destroying a context during registration at the' +
                ' point when the prototype\'s constructor property is retrieved.');
    jsTestIsAsync = true;
    successfullyParsed = true;
} else {
    // The child frame
    var proto = Object.create(HTMLElement.prototype, {
        constructor: {
            configurable: true,
            get: function () {
                destroyContext();
                return HTMLElement;
            }
        }
    });

    try {
        var element = document.registerElement('x-a', {prototype: proto});
        log('FAIL expected register to throw an exception');
    } catch (ex) {
        log('PASS caught expected exception "' + ex + '"');
    }
    done();
}
</script>
