third_party/WebKit/LayoutTests/fast/dom/custom/registration-context-delete-during-register-base-constructor-retrieval-expected.txt

Tests destroying a context during registration at the point when the prototype's constructor property is retrieved.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS destroyed context
PASS caught expected exception "InvalidStateError: Failed to execute 'registerElement' on 'Document': Registration failed for type 'x-a'. The context is no longer valid."
PASS child done
PASS successfullyParsed is true

TEST COMPLETE
